+++ packages/config-plugins/src/index.ts
@@ -57,5 +57,3 @@ export { withStaticPlugin } from './plugins/static-plugins';
 export { compileModsAsync } from './plugins/mod-compiler';
 
 export { PluginError } from './utils/errors';
-
-export * as BaseModPlugins from './plugins/compiler-plugins';
